+++ dd-sdk-android-gradle-plugin/src/main/kotlin/com/datadog/gradle/plugin/kcp/DatadogKotlinCompilerPluginSupport.kt
@@ -12,6 +12,7 @@ import org.gradle.api.logging.Logging
 import org.gradle.api.provider.Provider
 import org.jetbrains.kotlin.gradle.plugin.KotlinCompilation
 import org.jetbrains.kotlin.gradle.plugin.KotlinCompilerPluginSupportPlugin
+import org.jetbrains.kotlin.gradle.plugin.KotlinPlatformType
 import org.jetbrains.kotlin.gradle.plugin.SubpluginArtifact
 import org.jetbrains.kotlin.gradle.plugin.SubpluginOption
 import org.jetbrains.kotlin.gradle.plugin.getKotlinPluginVersion
@@ -19,9 +20,9 @@ import org.jetbrains.kotlin.gradle.plugin.getKotlinPluginVersion
 /**
  * Support class for configuring the Datadog Kotlin Compiler Plugin based on the Kotlin version.
  * This class implements [KotlinCompilerPluginSupportPlugin] to properly integrate with the
- * Kotlin Gradle Plugin and target the correct version-specific subplugin.
+ * Kotlin Gradle Plugin and target the correct version-specific sub-plugin.
  *
- * The plugin points to different subplugin based on kotlin compiler version (kotlin20, kotlin21, kotlin22).
+ * The plugin points to different sub-plugin based on kotlin compiler version (kotlin20, kotlin21, kotlin22, kotlin24).
  *
  */
 class DatadogKotlinCompilerPluginSupport : KotlinCompilerPluginSupportPlugin {
@@ -31,17 +32,26 @@ class DatadogKotlinCompilerPluginSupport : KotlinCompilerPluginSupportPlugin {
 
     @Suppress("ReturnCount")
     override fun isApplicable(kotlinCompilation: KotlinCompilation<*>): Boolean {
-        val project = kotlinCompilation.target.project
+        val kotlinTarget = kotlinCompilation.target
+        if (kotlinTarget.platformType != KotlinPlatformType.androidJvm) {
+            // TODO RUM-16076 Explore .common type support
+            LOGGER.info(
+                "Datadog Jetpack Compose instrumentation cannot be" +
+                    " applied to the Kotlin Compilation of platform type = ${kotlinTarget.platformType}"
+            )
+            return false
+        }
+        val project = kotlinTarget.project
         kgpVersion = project.getKotlinPluginVersion()
         kotlinVersion = KotlinVersion.from(kgpVersion)
         if (kotlinVersion == KotlinVersion.UNSUPPORTED) {
-            LOGGER.debug("Unsupported Kotlin version for the Datadog Jetpack Compose instrumentation: \$kgpVersion\"")
+            LOGGER.info("Unsupported Kotlin version for the Datadog Jetpack Compose instrumentation: \$kgpVersion\"")
             return false
         }
         val ddExtension = project.extensions.findByType(DdExtension::class.java) ?: return false
 
         if (ddExtension.composeInstrumentation == InstrumentationMode.DISABLE) {
-            LOGGER.debug("Datadog Jetpack Compose instrumentation is disabled")
+            LOGGER.info("Datadog Jetpack Compose instrumentation is disabled")
             return false
         }
 
@@ -54,6 +64,19 @@ class DatadogKotlinCompilerPluginSupport : KotlinCompilerPluginSupportPlugin {
         val project = kotlinCompilation.target.project
         val ddExtension = project.extensions.getByType(DdExtension::class.java)
         val composeInstrumentation = ddExtension.composeInstrumentation
+
+        // For Kotlin 2.4.0+, CompilerPluginRegistrar does not support LoadingOrder.FIRST,
+        // so we use -Xcompiler-plugin-order to guarantee Datadog runs before Compose.
+        // Use compileTaskProvider (KGP 2.0+ stable API) instead of the deprecated
+        // kotlinOptions.freeCompilerArgs which does not propagate reliably in KGP 2.4.0.
+        if (kotlinVersion == KotlinVersion.KOTLIN24) {
+            kotlinCompilation.compileTaskProvider.configure { task ->
+                task.compilerOptions.freeCompilerArgs.add(
+                    "-Xcompiler-plugin-order=$PLUGIN_ID>$COMPOSE_PLUGIN_ID"
+                )
+            }
+        }
+
         return project.provider {
             listOf(
                 SubpluginOption(
@@ -71,6 +94,7 @@ class DatadogKotlinCompilerPluginSupport : KotlinCompilerPluginSupportPlugin {
             KotlinVersion.KOTLIN19, KotlinVersion.KOTLIN20 -> "$PLUGIN_ARTIFACT_ID_PREFIX-kotlin20"
             KotlinVersion.KOTLIN21 -> "$PLUGIN_ARTIFACT_ID_PREFIX-kotlin21"
             KotlinVersion.KOTLIN22 -> "$PLUGIN_ARTIFACT_ID_PREFIX-kotlin22"
+            KotlinVersion.KOTLIN24 -> "$PLUGIN_ARTIFACT_ID_PREFIX-kotlin24"
             KotlinVersion.UNSUPPORTED -> error("Unsupported Kotlin version: $kgpVersion")
         }
         return SubpluginArtifact(
@@ -80,8 +104,9 @@ class DatadogKotlinCompilerPluginSupport : KotlinCompilerPluginSupportPlugin {
         )
     }
 
-    companion object {
+    private companion object {
         private const val PLUGIN_ID = "com.datadoghq.kotlin.compiler"
+        private const val COMPOSE_PLUGIN_ID = "androidx.compose.compiler.plugins.kotlin"
         private const val GROUP_ID = "com.datadoghq"
         private const val PLUGIN_ARTIFACT_ID_PREFIX = "dd-sdk-android-gradle-plugin-kcp"
         private const val INSTRUMENTATION_MODE = "INSTRUMENTATION_MODE"